RestaurantRequest, RestaurantCommand 수정 및 파일명 변경
