+++ src/main/java/com/happy/delivery/application/user/RestaurantSearchService.java
@@ -1,6 +1,6 @@
 package com.happy.delivery.application.user;
 
-import com.happy.delivery.application.common.restaurant.command.RestaurantCommand;
+import com.happy.delivery.application.common.restaurant.command.RestaurantSearchCommand;
 import com.happy.delivery.application.common.restaurant.result.RestaurantCategoryResult;
 import com.happy.delivery.application.common.restaurant.result.RestaurantResult;
 import java.util.List;
@@ -13,5 +13,5 @@ public interface RestaurantSearchService {
   List<RestaurantCategoryResult> getCategories();
 
   List<RestaurantResult> restaurantSearchByCategory(String category,
-      RestaurantCommand restaurantCommand);
+      RestaurantSearchCommand restaurantSearchCommand);
 }
